refactor(client): move client logic form core package to client
