+++ packages/client/src/actions/reconnect.ts
@@ -1,7 +1,10 @@
-import type { Address, Config, Connection, Connector } from '@bigmi/core'
+import type { Address } from '@bigmi/core'
 import type { ErrorType } from '@bigmi/core/src/errors/utils.js'
-import type { CreateConnectorFn } from '@bigmi/core/src/factories/createConnector.js'
+
 import type { Compute } from '@bigmi/core/src/types/utils.js'
+import type { Config } from '../core/createConfig.js'
+import type { Connection } from '../types/connection.js'
+import type { Connector, CreateConnectorFn } from '../types/connector.js'
 
 let isReconnecting = false
