+++ packages/client/src/actions/watchAccount.ts
@@ -1,4 +1,5 @@
-import { type Config, deepEqual } from '@bigmi/core'
+import { deepEqual } from '@bigmi/core'
+import type { Config } from '../core/createConfig.js'
 import { type GetAccountReturnType, getAccount } from './getAccount.js'
 
 export type WatchAccountParameters<config extends Config = Config> = {
