+++ packages/client/src/connectors/leather.ts
@@ -1,17 +1,16 @@
-import type {
-  Address,
-  BtcAccount,
-  SignPsbtParameters,
-  UTXOWalletProvider,
-} from '@bigmi/core'
+import type { Address, BtcAccount, SignPsbtParameters } from '@bigmi/core'
 import {
   MethodNotSupportedRpcError,
   ProviderNotFoundError,
   UserRejectedRequestError,
-  createConnector,
 } from '@bigmi/core'
 
-import type { ProviderRequestParams, UTXOConnectorParameters } from './types.js'
+import { createConnector } from '../core/createConnector.js'
+import type {
+  ProviderRequestParams,
+  UTXOConnectorParameters,
+  UTXOWalletProvider,
+} from './types.js'
 
 export type LeatherBitcoinEventMap = {
   accountChange(accounts: BtcAccount[]): void
