+++ packages/client/src/connectors/phantom.ts
@@ -1,16 +1,15 @@
-import type {
-  Address,
-  BtcAccount,
-  SignPsbtParameters,
-  UTXOWalletProvider,
-} from '@bigmi/core'
+import type { Address, BtcAccount, SignPsbtParameters } from '@bigmi/core'
 import {
   MethodNotSupportedRpcError,
   ProviderNotFoundError,
   UserRejectedRequestError,
-  createConnector,
 } from '@bigmi/core'
-import type { ProviderRequestParams, UTXOConnectorParameters } from './types.js'
+import { createConnector } from '../core/createConnector.js'
+import type {
+  ProviderRequestParams,
+  UTXOConnectorParameters,
+  UTXOWalletProvider,
+} from './types.js'
 
 export type PhantomBitcoinEventMap = {
   accountsChanged(accounts: BtcAccount[]): void
