+++ packages/client/src/connectors/xverse.ts
@@ -1,18 +1,17 @@
-import type {
-  Address,
-  BtcAccount,
-  SignPsbtParameters,
-  UTXOWalletProvider,
-} from '@bigmi/core'
+import type { Address, BtcAccount, SignPsbtParameters } from '@bigmi/core'
 import {
   MethodNotSupportedRpcError,
   ProviderNotFoundError,
   UserRejectedRequestError,
   base64ToHex,
-  createConnector,
   hexToBase64,
 } from '@bigmi/core'
-import type { ProviderRequestParams, UTXOConnectorParameters } from './types.js'
+import { createConnector } from '../core/createConnector.js'
+import type {
+  ProviderRequestParams,
+  UTXOConnectorParameters,
+  UTXOWalletProvider,
+} from './types.js'
 
 export type XverseBitcoinEventMap = {
   accountChange(accounts: BtcAccount[]): void
