+++ packages/client/src/core/createStorage.ts
@@ -1,36 +1,8 @@
-import type { PartializedState } from '../types/state.js'
-import type { Compute } from '../types/utils.js'
+import type { Compute } from '@bigmi/core'
+
+import type { Storage, StorageItemMap } from '../types/storage.js'
 import { deserialize as deserialize_ } from '../utils/deserialize.js'
 import { serialize as serialize_ } from '../utils/serialize.js'
-
-// key-values for loose autocomplete and typing
-export type StorageItemMap = {
-  recentConnectorId: string
-  state: PartializedState
-}
-
-export type Storage<
-  itemMap extends Record<string, unknown> = Record<string, unknown>,
-  ///
-  storageItemMap extends StorageItemMap = StorageItemMap & itemMap,
-> = {
-  key: string
-  getItem<
-    key extends keyof storageItemMap,
-    value extends storageItemMap[key],
-    defaultValue extends value | null | undefined,
-  >(
-    key: key,
-    defaultValue?: defaultValue | undefined
-  ):
-    | (defaultValue extends null ? value | null : value)
-    | Promise<defaultValue extends null ? value | null : value>
-  setItem<
-    key extends keyof storageItemMap,
-    value extends storageItemMap[key] | null,
-  >(key: key, value: value): void | Promise<void>
-  removeItem(key: keyof storageItemMap): void | Promise<void>
-}
 
 export type BaseStorage = {
   getItem(
