+++ packages/core/src/actions/signPsbt.ts
@@ -1,12 +1,12 @@
 import type { Transport } from '../factories/createTransport.js'
-import type { Account } from '../types/account.js'
-import type { Chain } from '../types/chain.js'
-import type { Client } from '../types/client.js'
 import type {
   SignPsbtParameters,
   SignPsbtReturnType,
   UTXOWalletSchema,
-} from '../types/client.js'
+} from '../transports/types.js'
+import type { Account } from '../types/account.js'
+import type { Chain } from '../types/chain.js'
+import type { Client } from '../types/client.js'
 
 export async function signPsbt<
   C extends Chain | undefined,
